Replace libsnmp with libnetsnmp, add /usr/local prefix (#124)
Allow the generator to compile under FreeBSD

Net-snmp library is at libnetsnmp.so, we use that instead of libsnmp.so.

Building the generator under FreeBSD requires that the /usr/local/include is searched for headers and the /usr/local/lib for libraries. Adding these two into the cgo CFLAGS and LDFLAGS ensures that Go under FreeBSD will be able to find the Net-SNMP library headers and shared objects.