snmp: Add ddwrt module
The list of SNMP OIDs to care about for DD-WRT can be found here:
https://www.dd-wrt.com/wiki/index.php/SNMP#Known_OID.C2.B4s_via_SNMP

This works with routers running DD-WRT. It includes the usual things
like network interfaces from SNMPv2-SMI and some information about the
device itself through HOST-RESOURCES-MIB. I left out support for
`hrSWRun`, `hrSWRunPerf`, `hrSWInstalled` and `hrMIBAdminInfo` as they
seem bad fits for what Prometheus would care about and in the case of my
device don't contain any data. Same thing for `hrDevice`.

The `1.3.6.1.4.1.2021` bits come from the UCD-SNMP-MIB, part of
Net-SNMP which is what appears to be the SNMPd on these devices:
* .10: laTable (load information)
* .11: systemStats