Update generator readme.md to include full make
Centos 7 has an issue with the generator parsing the default MIBs when using just 'make mibs'. Instead running the full 'make' command inside generator resolves the unmet dependency and allows the generator to run correctly. Without running full make the generator will always fail on generating MIBs, even when more are added in later as one of the defaults is incorrectly functioning.

Signed-off-by: Michael Rickert <racefpv@outlook.com>