Add support for Raritan PDUs.
This change downloads the latest PDU-MIB definition from Raritan,
which was last updated February 22, 2016, and then defines a handful
of useful values for monitoring.  This includes the current, voltage,
and wattage of both inlets and outlets on the PDU, and the current
switch status of each outlet.

This configuration has been tested and verified to work with a
Raritan DPXR8-15 unit in production.

Signed-off-by: Jeff Sharkey <jsharkey@jsharkey.org>