Update generator for net-snmp 5.9
At some point in the past few months, CircleCI updated the
circleci/golang:1.16 image from being based on Debian 10 (buster) to
Debian 11 (bullseye). With this, net-snmp (libsnmp-dev) was updated from
5.7.3 to 5.9.

In net-snmp 5.9, the definition of `tclist` (copied into the generator's
net_snmp.go) changed from

    #define MAXTC   4096
    struct tc {                     /* textual conventions */
        int             type;
        int             modid;
        char           *descriptor;
        char           *hint;
        struct enum_list *enums;
        struct range_list *ranges;
        char           *description;
    } tclist[MAXTC];

to

    #define TC_INCR 100
    struct tc {                     /* textual conventions */
        int             type;
        int             modid;
        char           *descriptor;
        char           *hint;
        struct enum_list *enums;
        struct range_list *ranges;
        char           *description;
    } *tclist;
    int tc_alloc;

The mismatch between the definitions used by net-snmp and snmp_exporter
caused the removal of all `fixed_size` fields in the generated snmp.yml,
which in turn caused CI runs for recent PRs to fail (some examples of
such PRs are 702, 700, 681).

This commit adapts net_snmp.go to the new definition of `tclist`.

Signed-off-by: Torbjörn Lönnemark <tobbez@ryara.net>