config: allow setting default auth via YAML using `default: true`
Adds support for marking an SNMP auth block as the default by specifying `default: true`
in the YAML configuration. When present, the marked auth is assigned to `DefaultAuth`
and used wherever a default is needed. If multiple auths are marked as default, an error is returned.

This makes the default behavior configurable without code changes.